not sure what happened here, my bad my future self

+++ Srcs/Lists/ft_lstclear.c
@@ -6,13 +6,13 @@
 /*   By: amarini- <amarini-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/15 10:14:10 by amarini-          #+#    #+#             */
-/*   Updated: 2021/06/16 14:24:18 by amarini-         ###   ########.fr       */
+/*   Updated: 2021/06/28 12:19:43 by amarini-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../Includes/libft.h"
 
-void	ft_lstclear(t_list **lst, void (*del)(int *))
+void	ft_lstclear(t_list **lst, void (*del)(long long int *))
 {
 	t_list	*current;
 	t_list	*tmp;
