+++ Srcs/Lists/ft_lstdelone.c
@@ -6,13 +6,13 @@
 /*   By: amarini- <amarini-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/15 10:12:43 by amarini-          #+#    #+#             */
-/*   Updated: 2021/06/16 16:18:10 by amarini-         ###   ########.fr       */
+/*   Updated: 2021/06/28 12:20:52 by amarini-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../Includes/libft.h"
 
-void	ft_lstdelone(t_list *lst, void (*del)(int *))
+void	ft_lstdelone(t_list *lst, void (*del)(long long int *))
 {
 	if (!lst)
 		return ;
