+++ Srcs/Lists/ft_lstiter.c
@@ -6,13 +6,13 @@
 /*   By: amarini- <amarini-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/15 10:15:49 by amarini-          #+#    #+#             */
-/*   Updated: 2021/06/16 14:23:46 by amarini-         ###   ########.fr       */
+/*   Updated: 2021/06/28 12:20:58 by amarini-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../Includes/libft.h"
 
-void	ft_lstiter(t_list *lst, void (*f)(int *))
+void	ft_lstiter(t_list *lst, void (*f)(long long int *))
 {
 	if (!lst)
 		return ;
